+++ CHANGELOG.md
@@ -860,4 +860,5 @@
 - 优化 ResvgJNI 渲染方案；
 - 优化 WebDriver 渲染方案；
 - 优化 ResourceLoader 组件；
+- 弃用 HtmlTemplateUtil 和 模板构建工具；
 - 优化 部分其他代码。